Disable registration of shared rack_monitor in main container
Since we have a rack_monitor in the top-level umbrella container, and that’s attached to the corresponding top-level `Blog::Application` rack app, this is what will take care of all our incoming request logging. We don’t need to log the requests a second time once they’re dispatched to the `Main::Application` sub-app.