Adjust sub-app nesting and naming to match dry-web-roda convention
As of version 0.9, dry-web-roda generates sub-apps with classes nested within the main app’s top-level module. This reduces global constant pollution, makes it clearer the sub-app is a constituent part of the overall app, and makes constant resolution more straightforward from within sub-apps (when trying to access constants defined within the app’s top-level module).

In this change, also move the article repo outside of its previous “Persistence” module, since it’s not actually part of the persistence subsystem itself, but rather an interface to it.